requirements.txt
opencv-python
streamlit
google-genai
python-dotenv